requirements.txt
aiogram==3.6.0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
tzlocal==5.2
openpyxl==3.1.5
numpy==1.26.4
pandas==2.2.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
